Resolved Race Conditions and SD Error
Resolved race condition when auxiliary buffer is full, as well as multiple issues with how the auxiliary buffer was fed into the primary Influx Client buffer. Also fixed an issue with the file path for the SD causing no file to be created. Added additional mode to only store datapoints that got a write error from Influx to the SD card.